pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tmiplus"
version = "0.1.0"
description = "TMI Resource Planning CLI (prototype)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "TMI", email = "noreply@example.com" }]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pydantic-core>=2.18.0",
    "python-dateutil>=2.8.2",
    "dateparser>=1.1.8",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
    "click>=8.1.7",
    "pyairtable>=2.2.3",
]

[project.optional-dependencies]
ilp = [
    "pulp>=2.7.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
    "black>=24.4.2",
    "nox>=2024.4.15",
]

[project.scripts]
tmi = "tmiplus.tli.__main__:app"
